+++ example/lib/src/bloc/default_bloc.dart
@@ -6,12 +6,12 @@ part 'default_event.dart';
 part 'default_state.dart';
 
 class DefaultBloc extends Bloc<DefaultEvent, DefaultState> {
-  final AuthRepository _authRepository;
+  final IAuthRepository _authRepository;
 
   DefaultBloc(this._authRepository) : super(DefaultState()) {
-    on<DefaultEvent>((event, emit) {
-      _authRepository.login();
-      emit(DefaultState(authorized: true));
+    on<DefaultEvent>((event, emit) async {
+      final result = await _authRepository.login();
+      emit(DefaultState(authorized: true, token: result));
     });
   }
 }