+++ example/lib/src/bloc/default_state.dart
@@ -2,6 +2,7 @@ part of 'default_bloc.dart';
 
 class DefaultState {
   final bool authorized;
+  final String? token;
 
-  DefaultState({this.authorized = false});
+  DefaultState({this.authorized = false, this.token});
 }