+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 5.0.1
+
+* DependencyFactory method create return FutureOr
+
+## 5.0.1-dev
+
+* DependencyFactory method create return FutureOr
+
 ## 5.0.0
 
 * DependencyFactory for initialize DependencyContainer